% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previews.R
\name{preview_uniques}
\alias{preview_uniques}
\title{List unique values for all columns.}
\usage{
preview_uniques(wqp)
}
\arguments{
\item{wqp}{A dataframe from the WQP datawarehouse.}
}
\value{
A dataframe of unique column values.
}
\description{
List unique values for all columns.
}
\examples{
preview_uniques(wqp)
}
